requirements.txt
appnope==0.1.0
astroid==2.3.2
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
coveralls==2.0.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.6
django-dotenv==1.4.2
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
entrypoints==0.3
flake8==3.7.9
idna==2.8
importlib-metadata==1.4.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
more-itertools==8.1.0
oauthlib==3.1.0
packaging==20.1
parso==0.7.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.3
pyparsing==2.4.6
pytest==5.3.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
verification-code==1.0.0
wcwidth==0.1.8
wrapt==1.11.2
zipp==2.0.1
